rs/requirements.txt
absl-py==0.9.0
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
astunparse==1.6.3
async-timeout==4.0.1
attrs==21.2.0
azure-functions==1.2.1
backcall==0.1.0
bcrypt==3.2.0
bottle==0.12.18
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
charset-normalizer==2.0.9
click==7.1.2
cryptography==2.9.2
decorator==4.4.2
Flask==1.1.2
flatbuffers==2.0
frozenlist==1.2.0
fsspec==2021.11.1
gast==0.3.3
gcsfs==2021.11.1
google-api-core==2.3.0
google-auth==2.3.3
google-auth-oauthlib==0.4.1
google-cloud==0.34.0
google-cloud-core==2.2.1
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.28.1
h5py==2.10.0
httptools==0.3.0
idna==2.9
ipython==7.14.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.14.1
keras==2.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
libclang==12.0.0
Markdown==3.2.1
MarkupSafe==1.1.1
multidict==5.2.0
mysql-connector-python==8.0.20
nltk==3.5
numpy==1.18.4
oauthlib==3.1.0
opt-einsum==3.2.1
pandas==1.0.3
paramiko==2.8.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
protobuf==3.19.1
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyNaCl==1.4.0
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.5.7
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
sanic==21.9.3
sanic-routing==0.7.2
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
storage==0.0.4.3
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.22.0
tensorflow-text==2.7.3
termcolor==1.1.0
tqdm==4.46.0
traitlets==4.3.3
typing-extensions==4.0.1
ujson==4.3.0
urllib3==1.25.9
uvloop==0.16.0
uWSGI==2.0.18
wcwidth==0.1.9
websockets==10.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.7.2
